backend/pyproject.toml
[tool.poetry]
name = "cursor-excel-ai-backend"
version = "0.1.0"
description = "AI-powered Excel assistant backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
langgraph = "^0.2.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
openai = "^1.6.1"
pandas = "^2.1.4"
pydantic = "^2.5.2"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 